feat(passengerDetails): add validation
